arm64: ssbd: Introduce thread flag to control userspace mitigation
Source: linux-mvista-2.4
MR: 94992, 00000
Type: Integration
Disposition: Merged from linux-mvista-2.4
ChangeID: e7d02797288f12ca257cdab49d19ccf9b3d2cb12
Description:

commit 9dd9614f5476687abbff8d4b12cd08ae70d7c2ad upstream.

In order to allow userspace to be mitigated on demand, let's
introduce a new thread flag that prevents the mitigation from
being turned off when exiting to userspace, and doesn't turn
it on on entry into the kernel (with the assumption that the
mitigation is always enabled in the kernel itself).

This will be used by a prctl interface introduced in a later
patch.

Reviewed-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Corey Minyard <cminyard@mvista.com>
Signed-off-by: Jeremy Puhlman <jpuhlman@mvista.com>